WIP splitting StreamProcessor & subscription state
