+++ Source/Events/Store/Actors/StreamProcessorStateManager.cs
@@ -24,12 +24,9 @@ public class StreamProcessorStateManager : StreamProcessorStateBase
 
 
     readonly Dictionary<StreamProcessorId, Bucket> _processorStates = new();
-    readonly Dictionary<StreamProcessorId, Bucket> _changedProcessorStates = new();
-    
-    readonly Dictionary<ScopeId, Dictionary<StreamSubscriptionId, Bucket>> _subscriptionStates = new();
-    readonly Dictionary<ScopeId, Dictionary<StreamSubscriptionId, Bucket>> _changedSubscriptionStates = new();
-    
-    readonly HashSet<ScopeId> _activeRequests = new();
+    Dictionary<StreamProcessorId, Bucket> _changedProcessorStates = new();
+
+    bool _activeRequest = false;
     readonly Dictionary<StreamSubscriptionId, Task<Bucket>> _waitingForSubscriptionStates = new();
 
 
@@ -63,7 +60,7 @@ public class StreamProcessorStateManager : StreamProcessorStateBase
         Context.ReenterAfter(_shutdownHook.ShuttingDown, () =>
         {
             _shuttingDown = true;
-            if (_activeRequests.Count == 0) // No current changes
+            if (!_activeRequest) // No current changes
             {
                 _shutdownHook.MarkCompleted();
             }
@@ -90,70 +87,6 @@ public class StreamProcessorStateManager : StreamProcessorStateBase
         return Task.FromResult(response);
     }
 
-    public override Task GetBySubscriptionId(StreamSubscriptionId subscriptionId, Action<StreamProcessorStateResponse> respond, Action<string> onError)
-    {
-        if (_shuttingDown)
-        {
-            onError("Runtime is shutting down");
-            return Task.CompletedTask;
-        }
-
-        if (!_waitingForSubscriptionStates.TryGetValue(subscriptionId, out var resultTask))
-        {
-            resultTask = LoadSubscription(subscriptionId);
-        }
-
-        Context.ReenterAfter(resultTask, _ =>
-        {
-            if (_.IsCompletedSuccessfully)
-            {
-                respond(new StreamProcessorStateResponse
-                {
-                    StreamKey = new StreamProcessorKey
-                    {
-                        SubscriptionId = subscriptionId
-                    },
-                    Bucket = { resultTask!.Result }
-                });
-            }
-            else
-            {
-                onError("Failed to get subscription state: " + resultTask.Exception?.Message);
-            }
-        });
-        return Task.CompletedTask;
-    }
-
-
-    Task<Bucket> LoadSubscription(StreamSubscriptionId subscriptionId)
-    {
-        var taskCompletionSource = new TaskCompletionSource<Bucket>();
-        var task = _repository.TryGet(SubscriptionId.FromProtobuf(subscriptionId), Context.CancellationToken);
-        _waitingForSubscriptionStates.Add(subscriptionId, taskCompletionSource.Task);
-        Context.ReenterAfter(task, _ =>
-        {
-            try
-            {
-                if (task is { IsCompletedSuccessfully: true, Result.Success: true })
-                {
-                    taskCompletionSource.SetResult(task.Result.Result.ToProtobuf());
-                }
-                else
-                {
-                    if (task.Exception is not null)
-                    {
-                        taskCompletionSource.SetException(task.Exception);
-                    }
-                }
-            }
-            finally
-            {
-                _waitingForSubscriptionStates.Remove(subscriptionId);
-            }
-        });
-        return taskCompletionSource.Task;
-    }
-
     public override Task SetByProcessorId(SetStreamProcessorStateRequest request)
     {
         if (request.StreamKey.IdCase != StreamProcessorKey.IdOneofCase.StreamProcessorId)
@@ -167,60 +100,23 @@ public class StreamProcessorStateManager : StreamProcessorStateBase
         return Task.CompletedTask;
     }
 
-    public override Task SetBySubscriptionId(SetStreamProcessorStateRequest request)
+    void PersistCurrentState()
     {
-        if (request.StreamKey.IdCase != StreamProcessorKey.IdOneofCase.SubscriptionId)
-        {
-            throw new ArgumentException("Can only set subscription state by subscription id");
-        }
-        var subscriptionId = request.StreamKey.SubscriptionId;
-        ScopeId scopeId = subscriptionId.ScopeId.ToGuid();
-
-        SetSubscriptionState(scopeId, subscriptionId, request.Bucket);
-        AddChange(scopeId, subscriptionId, request.Bucket);
-        
-        PersistCurrentSubscriptionState(scopeId);
-        return Task.CompletedTask;
-    }
-
-    void SetSubscriptionState(ScopeId scope, StreamSubscriptionId subscriptionId, Bucket state)
-    {
-        if (!_subscriptionStates.TryGetValue(scope, out var scopedSubscriptionStates))
-        {
-            _subscriptionStates.Add(scope, scopedSubscriptionStates = new());
-        }
-
-        scopedSubscriptionStates[subscriptionId] = state;
-    }
-    
-    void AddChange(ScopeId scope, StreamSubscriptionId subscriptionId, Bucket state)
-    {
-        if (!_changedSubscriptionStates.TryGetValue(scope, out var changedScopedSubscriptionStates))
-        {
-            _changedSubscriptionStates.Add(scope, changedScopedSubscriptionStates = new());
-        }
-        
-        changedScopedSubscriptionStates[subscriptionId] = state;
-    }
-
-    void PersistCurrentSubscriptionState(ScopeId scopeId)
-    {
-        if (_activeRequests.Contains(scopeId))
+        if (_activeRequest)
         {
             return;
         }
 
-        if (_changedSubscriptionStates.Count == 0) return;
-        if (!_changedSubscriptionStates.Remove(scopeId, out var currentChanges))
-        {
-            return; // No changes for this scope
-        }
+        if (_changedProcessorStates.Count == 0) return;
+        var currentChanges = _changedProcessorStates;
+        _changedProcessorStates = new Dictionary<StreamProcessorId, Bucket>();
 
-        _activeRequests.Add(scopeId);
-        // Persist(currentChanges, scopeId);
+
+        _activeRequest = true;
+        // Persist(currentChanges);
     }
 
-    // void Persist(Dictionary<SubscriptionId, Bucket> changes, ScopeId scopeId)
+    // void Persist(Dictionary<StreamProcessorId, Bucket> changes)
     // {
     //     var streamProcessorStates = changes.ToDictionary(_ => _.Key.FromProtobuf(), _ => _.Value.FromProtobuf());
     //     var persistTask = _repository.Persist(streamProcessorStates, Context.CancellationToken);
@@ -228,18 +124,18 @@ public class StreamProcessorStateManager : StreamProcessorStateBase
     //     {
     //         if (persistTask.IsCompletedSuccessfully)
     //         {
-    //             _activeRequests.Remove(scopeId);
-    //             PersistCurrentSubscriptionState(scopeId);
-    //             if (_shuttingDown && _activeRequests.Count == 0)
+    //             _activeRequest = false;
+    //             PersistCurrentState();
+    //             if (_shuttingDown && !_activeRequest)
     //             {
     //                 _shutdownHook.MarkCompleted();
     //             }
     //         }
     //         else
     //         {
-    //             _logger.FailedToPersistStreamProcessorState(persistTask.Exception!, scopeId);
+    //             _logger.FailedToPersistStreamProcessorState(persistTask.Exception!);
     //             // Try again
-    //             Persist(changes, scopeId);
+    //             Persist(changes);
     //         }
     //     });
     // }
